requirements.txt
uvicorn~=0.30.6
fastapi~=0.112.2
pydantic~=2.8.2
requests~=2.32.3
six~=1.16.0
python-dotenv~=1.0.1